+++ test/spectests/selftests.cc
@@ -16,8 +16,8 @@ TEST(Prefix, SizeCheck)
 
 TEST(Files, SelfTest)
 {
-    EXPECT_NE(0, magicNum1);
-    EXPECT_EQ(0, magicNum1 / 0xFF);
+    EXPECT_NE(static_cast<size_t>(0), magicNum1);
+    EXPECT_EQ(static_cast<size_t>(0), magicNum1 / 0xFF);
     struct stat stats;
     EXPECT_EQ(0, stat(zeroFile.c_str(), &stats));
     EXPECT_EQ(0, stat(smallFile.c_str(), &stats));
